% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verbs.R
\name{Analyze}
\alias{Analyze}
\title{Create an analysis file}
\usage{
Analyze(fileStem, inputFiles, title)
}
\arguments{
\item{fileStem}{The file stem}

\item{inputFiles}{A vector of files to load}

\item{title}{Title of the document}
}
\description{
Create an analysis file
}

