% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verbs.R
\name{Cook}
\alias{Cook}
\title{Create a cook file}
\usage{
Cook(fileStem, inputFiles, title)
}
\arguments{
\item{fileStem}{The file stem}

\item{inputFiles}{A vector of files to load}

\item{title}{Title of the Cook document}
}
\description{
Create a cook file
}

