% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\name{DescribeObjects}
\alias{DescribeObjects}
\title{Describe the objects}
\usage{
DescribeObjects(objects, FUNS)
}
\arguments{
\item{objects}{The objects to be described}

\item{FUNS}{the functions which will describe the objects}
}
\description{
Describe the objects
}

