% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verbs.R
\name{Gather}
\alias{Gather}
\title{Create a gather file}
\usage{
Gather(fileStem, title)
}
\arguments{
\item{fileStem}{The file stem}

\item{title}{Title of the Gather document}

\item{inputFiles}{A vector of files to load}
}
\description{
Create a gather file
}

