% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\name{GatherObjects}
\alias{GatherObjects}
\title{Gather objects}
\usage{
GatherObjects(patterns = c("^df", "^plt", "^fit"), env)
}
\arguments{
\item{patterns}{A vector of patterns}

\item{env}{An enviroment. If missing, the global environment will be used.}
}
\description{
Gather objects
}

