% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{GetSQL}
\alias{GetSQL}
\title{Read a SQL file}
\usage{
GetSQL(filename, warn = FALSE)
}
\arguments{
\item{filename}{A character variable with a filename}

\item{warn}{Warn if there is a missing final end of line. Default is TRUE}
}
\value{
String
}
\description{
Read a SQL file
}
