% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\name{ListObjects}
\alias{ListObjects}
\title{List objects}
\usage{
ListObjects(patterns = c("^df", "^plt", "^fit"), env)
}
\arguments{
\item{patterns}{A vector of patterns}

\item{env}{An enviroment. If missing, the global environment will be used.}
}
\value{
character vector
}
\description{
Form a character vector of the names of objects, based on patterns.
}
\details{
This will collect objects based on a vector of character patterns.
}

