% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{LoadInput}
\alias{LoadInput}
\title{LoadInput}
\usage{
LoadInput(params)
}
\arguments{
\item{params}{A list of parameters. Should include the named items inputDir and inputfiles}
}
\value{
Character vector of names of loaded objects
}
\description{
Load files from a defined input directory and set of files
}

