% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Objects.R
\name{LoadObjects}
\alias{LoadObjects}
\title{LoadObjects}
\usage{
LoadObjects(params, loadFiles)
}
\arguments{
\item{params}{A list of parameters. Should include the named items inputDir and inputFiles}

\item{loadFiles}{A list of}
}
\value{
Character vector of names of loaded objects
}
\description{
Load objects from a defined input directory and set of files.
Objects are loaded into the Global environment.
}

