% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.R
\name{Make}
\alias{Make}
\title{Create a new analysis}
\usage{
Make(path)
}
\arguments{
\item{path}{A character variable indicating the name of the project to create}

\item{packrat}{Boolean indicating whether or not packrat should be used. Default is TRUE.}

\item{RStudio}{Boolean indicating whether or not to create an RStudio project. Default is TRUE.}

\item{git}{Boolean indicating whether to initialize a new git repository. Default is TRUE.}

\item{overwrite}{Overwrite the directory if it already exists. Default is FALSE.}

\item{createNewDirectory}{Create a new directory? Default is TRUE.}
}
\value{
NewAnalysis will invisibly return the name of the project.
}
\description{
Create a new analysis
}

