% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\name{NamesToObjects}
\alias{NamesToObjects}
\title{NamesToObjects}
\usage{
NamesToObjects(x, envir)
}
\arguments{
\item{x}{Vector or list of named objects}

\item{envir}{Environment in which to search}
}
\value{
A list of objects
}
\description{
Returns a list of the objects named in x
}
\details{
Return a list of named objects
}

