% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewResearch.R
\name{NewResearch}
\alias{NewResearch}
\title{Create a new Research}
\usage{
NewResearch(path = ".", RStudio = TRUE, git = TRUE, packrat = TRUE,
  overwrite = FALSE, createNewDirectory = TRUE)
}
\arguments{
\item{path}{A character variable indicating the name of the project to create}

\item{RStudio}{Boolean indicating whether or not to create an RStudio project. Default is TRUE.}

\item{git}{Boolean indicating whether to initialize a new git repository. Default is TRUE.}

\item{packrat}{Boolean indicating whether or not packrat should be used. Default is TRUE.}

\item{overwrite}{Overwrite the directory if it already exists. Default is FALSE.}

\item{createNewDirectory}{Create a new directory? Default is TRUE.}
}
\value{
NewResearch will invisibly return the name of the project.
}
\description{
Create a new Research
}

