% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputFile.R
\name{OutputFile}
\alias{OutputFile}
\title{OutputFile}
\usage{
OutputFile(params)
}
\arguments{
\item{params}{List object with output directory and file stem elements}
}
\description{
Construct an output file from a params list.
}
