% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verbs.R
\name{Present}
\alias{Present}
\title{Create a presentation file}
\usage{
Present(fileStem, inputFiles, title, outputFormat)
}
\arguments{
\item{fileStem}{The file stem}

\item{inputFiles}{A vector of files to load}

\item{title}{Title of the Cook document}

\item{outputFormat}{outputFormat}
}
\description{
Create a presentation file
}

