% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\name{DescribeObjects}
\alias{DescribeObjects}
\title{Describe the objects}
\usage{
DescribeObjects(objects, FUNS)
}
\arguments{
\item{objects}{A list of objects to be described}

\item{FUNS}{the functions which will describe the objects}
}
\description{
This will apply functions to a vector of objects.
}
\details{
This may be used to quickly summarize a set of objects.

x may be given as a character vector or a list.
}
\examples{
\dontrun{
myFit1 <- lm(myData, y ~ x1 + x2)
myFit2 <- lm(myData, y ~ x3)

DesribeObjects(list(myFit1, myFit2), broom)
}

}

