% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResearchDirs.R
\name{NewResearch}
\alias{NewResearch}
\title{Create a new Research}
\usage{
NewResearch(path = ".", RStudio = TRUE, useExistingDirectory = FALSE)
}
\arguments{
\item{path}{A character variable indicating the name of the project to create}

\item{RStudio}{Boolean indicating whether or not to create an RStudio project. Default is TRUE.}

\item{useExistingDirectory}{Overwrite the directory if it already exists. Default is FALSE.}
}
\value{
NewResearch will invisibly return the name of the project.
}
\description{
Create a new Research
}

